CubeBtn/GetImage_M5stack/requirements.txt
google-api-python-client
oauth2client
opencv-python
numpy
Pillow